LayoutTests/fast/text/variations/optical-sizing-expected.txt

This test makes sure that font-optical-sizing works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(document.getElementById('test1')).getPropertyValue('font-optical-sizing') is "auto"
PASS window.getComputedStyle(document.getElementById('test2')).getPropertyValue('font-optical-sizing') is "auto"
PASS window.getComputedStyle(document.getElementById('test3')).getPropertyValue('font-optical-sizing') is "auto"
PASS window.getComputedStyle(document.getElementById('test4')).getPropertyValue('font-optical-sizing') is "none"
PASS window.getComputedStyle(document.getElementById('test5')).getPropertyValue('font-optical-sizing') is "none"
PASS document.getElementById('test1').style.fontOpticalSizing is ""
PASS document.getElementById('test2').style.fontOpticalSizing is "initial"
PASS document.getElementById('test3').style.fontOpticalSizing is "auto"
PASS document.getElementById('test4').style.fontOpticalSizing is "none"
PASS document.getElementById('test5').style.fontOpticalSizing is "inherit"
PASS window.getComputedStyle(document.getElementById('test1')).fontOpticalSizing is "auto"
PASS window.getComputedStyle(document.getElementById('test2')).fontOpticalSizing is "auto"
PASS window.getComputedStyle(document.getElementById('test3')).fontOpticalSizing is "auto"
PASS window.getComputedStyle(document.getElementById('test4')).fontOpticalSizing is "none"
PASS window.getComputedStyle(document.getElementById('test5')).fontOpticalSizing is "none"
PASS Found optical sizing
PASS successfullyParsed is true

TEST COMPLETE
Hello
Hello
Hello
Hello
Hello
